requirements.txt
hydra-core
omegaconf
torch
numpy
matplotlib
wandb
